XXX/requirements.txt
supabase>=2.4.0
openai>=1.35.3
python-dotenv>=1.0.1
numpy>=1.26.0
streamlit>=1.28.2
pandas>=2.0.3
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0